server/requirements.txt
Flask==2.2.3
fastapi==0.97.0
uvicorn==0.22.0
langchain==0.0.242
numpy==1.23.4
openai==0.27.8
PyMuPDF==1.22.3
python-dotenv==1.0.0
supabase==1.0.3
tqdm==4.64.1
